more columns in bookings

+++ database/migrations/2020_08_01_133326_create_sales_order_returns_table.php
@@ -15,8 +15,8 @@ class CreateSalesOrderReturnsTable extends Migration
     {
         Schema::create('sales_order_returns', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('booking_id');
-            $table->unsignedInteger('bookeditem_id');
+            $table->unsignedBigInteger('booking_id');
+            $table->unsignedBigInteger('bookeditem_id');
             $table->unsignedInteger('producto_id');
             $table->string('return_token', 50);
             $table->decimal('return_qty',9,2);
